add typeid_t ref-in size_t add bytearray_*

+++ include/pybind/adapter/constructor_adapter.hpp
@@ -11,11 +11,11 @@ namespace pybind
         : public adapter_interface
     {
     public:
-        constructor_adapter_interface( uint32_t _arity );
+        constructor_adapter_interface( size_t _arity );
         ~constructor_adapter_interface() override;
 
     public:
-        uint32_t arity() const;
+        size_t arity() const;
 
     public:
         virtual void * call( kernel_interface * _kernel, PyObject * _obj, PyObject * _args ) = 0;
@@ -26,7 +26,7 @@ namespace pybind
 #endif
 
     protected:
-        uint32_t m_arity;
+        size_t m_arity;
     };
     //////////////////////////////////////////////////////////////////////////
     typedef intrusive_ptr<constructor_adapter_interface> constructor_adapter_interface_ptr;
