+++ include/pybind/call/function_call.hpp
@@ -12,8 +12,8 @@ namespace pybind
     template<class F, class Ret>
     struct function_call_impl
     {
-        template<uint32_t ... I>
-        static Ret call( kernel_interface * _kernel, F f, PyObject * _arg, std::integer_sequence<uint32_t, I...> )
+        template<size_t ... I>
+        static Ret call( kernel_interface * _kernel, F f, PyObject * _arg, std::integer_sequence<size_t, I...> )
         {
             return (*f)(
                 tuple_getitem_t( _kernel, _arg, I ) ...
@@ -21,23 +21,23 @@ namespace pybind
         }
     };
 
-    template<class F, uint32_t Arity, class Ret>
+    template<class F, size_t Arity, class Ret>
     struct function_call_ret_impl
     {
         static PyObject * call( kernel_interface * _kernel, F f, PyObject * _arg )
         {
-            PyObject * py_result = detail::return_operator_t( _kernel, function_call_impl<F, Ret>::call( _kernel, f, _arg, std::make_integer_sequence<uint32_t, Arity>() ) );
+            PyObject * py_result = detail::return_operator_t( _kernel, function_call_impl<F, Ret>::call( _kernel, f, _arg, std::make_integer_sequence<size_t, Arity>() ) );
 
             return py_result;
         }
     };
 
-    template<class F, uint32_t Arity>
+    template<class F, size_t Arity>
     struct function_call_ret_impl<F, Arity, void>
     {
         static PyObject * call( kernel_interface * _kernel, F f, PyObject * _arg )
         {
-            function_call_impl<F, void>::call( _kernel, f, _arg, std::make_integer_sequence<uint32_t, Arity>() );
+            function_call_impl<F, void>::call( _kernel, f, _arg, std::make_integer_sequence<size_t, Arity>() );
 
             return _kernel->ret_none();
         }
@@ -54,8 +54,8 @@ namespace pybind
             static_assert(std::is_same<typename f_info::template reverse_iterator_param<0>, const pybind::args &>::value == false, "[pybind] use args bind");
 
 #if defined(PYBIND_DEBUG)
-            uint32_t arg_size = (_arg != nullptr) ? _kernel->tuple_size( _arg ) : 0;
-            uint32_t fn_arity = f_info::arity;
+            size_t arg_size = (_arg != nullptr) ? _kernel->tuple_size( _arg ) : 0;
+            size_t fn_arity = f_info::arity;
 
             if( arg_size != fn_arity )
             {
