+++ include/pybind/call/function_proxy_kernel_args_call.hpp
@@ -14,8 +14,8 @@ namespace pybind
     template<class P, class F, class Ret>
     struct function_proxy_kernel_args_call_impl
     {
-        template<uint32_t ... I>
-        static Ret call( kernel_interface * _kernel, P * _proxy, F f, PyObject * _arg, std::integer_sequence<uint32_t, I...> )
+        template<size_t ... I>
+        static Ret call( kernel_interface * _kernel, P * _proxy, F f, PyObject * _arg, std::integer_sequence<size_t, I...> )
         {
             return (*f)(_kernel, _proxy
                 , tuple_getitem_t( _kernel, _arg, I ) ...
@@ -24,23 +24,23 @@ namespace pybind
         }
     };
 
-    template<class P, class F, uint32_t Arity, class Ret>
+    template<class P, class F, size_t Arity, class Ret>
     struct function_proxy_kernel_args_call_ret_impl
     {
         static PyObject * call( kernel_interface * _kernel, P * _proxy, F f, PyObject * _arg )
         {
-            PyObject * py_result = detail::return_operator_t( _kernel, function_proxy_kernel_args_call_impl<P, F, Ret>::call( _kernel, _proxy, f, _arg, std::make_integer_sequence<uint32_t, Arity>() ) );
+            PyObject * py_result = detail::return_operator_t( _kernel, function_proxy_kernel_args_call_impl<P, F, Ret>::call( _kernel, _proxy, f, _arg, std::make_integer_sequence<size_t, Arity>() ) );
 
             return py_result;
         }
     };
 
-    template<class P, class F, uint32_t Arity>
+    template<class P, class F, size_t Arity>
     struct function_proxy_kernel_args_call_ret_impl<P, F, Arity, void>
     {
         static PyObject * call( kernel_interface * _kernel, P * _proxy, F f, PyObject * _arg )
         {
-            function_proxy_kernel_args_call_impl<P, F, void>::call( _kernel, _proxy, f, _arg, std::make_integer_sequence<uint32_t, Arity>() );
+            function_proxy_kernel_args_call_impl<P, F, void>::call( _kernel, _proxy, f, _arg, std::make_integer_sequence<size_t, Arity>() );
 
             return _kernel->ret_none();
         }
@@ -57,8 +57,8 @@ namespace pybind
             static_assert(std::is_same<typename f_info::template reverse_iterator_param<0>, const pybind::args &>::value == true, "[pybind] use args bind");
 
 #if defined(PYBIND_DEBUG)
-            uint32_t arg_size = _kernel->tuple_size( _arg );
-            uint32_t fn_arity = f_info::arity;
+            size_t arg_size = _kernel->tuple_size( _arg );
+            size_t fn_arity = f_info::arity;
 
             if( arg_size + 3 != fn_arity )
             {
